:recycle: refactor: move mysql files into separate directory
