+++ src/mysql/.env.development
@@ -1,0 +1,8 @@
+# This file is included in version control
+# Make sure to never include secrets in this file
+# They should be added to `.env.development.local` instead.
+DB_HOST=localhost
+DB_USER=mysql
+DB_PASSWORD=password
+DB_DATABASE=mysql
+DB_PORT=3306
